made primes a commandline argument
instead of having the file name and worksize static in the Node, the amount of primes is now a command line argument. To use it, set -pr to 100/1000/10000/100000, the worksize is dynamic and determined by the amount of primes.

For this, setters for PrimesFile and WorkSize were added.

Moved the definition of the amount of primes in the client to command line argument -pr.